CreateForm->form2XML->XML2form (basic) done
The form data is sent with the saveForm function in saveForm.js to the
form2XML.php script. This script creates an XML file and saves it to the
server. The XML2form.php script loads this XML (fetches the file based
on a GET submitted value with the id of the form) and outputs its
content in HTML.